Summary method for sfclust objects

This function summarizes the cluster assignments from the desired clustering sample.

Usage

# S3 method for class 'sfclust'
summary(object, sample = object$clust$id, sort = FALSE, ...)

Arguments

object

An object of class 'sfclust'.

sample

An integer specifying the clustering sample number to be summarized (default is the last sample).

sort

Logical value indicating if clusters should be relabel based on number of elements.

...

Additional arguments passed to print.default.